#23 - Separated validation rules as classes
