+++ app/Http/Requests/User/RegisterUserRequest.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Brewmap\Http\Requests\User;
+
+use Brewmap\Http\Requests\User\Rules\EmailRules;
+use Brewmap\Http\Requests\User\Rules\PasswordRules;
+use Brewmap\Http\Requests\BaseRequest;
+use Brewmap\Http\Requests\User\Rules\NameRules;
+
+class RegisterUserRequest extends BaseRequest
+{
+    public function rules(): array
+    {       
+        return [
+            "email" => EmailRules::rules([
+                "unique:users"
+            ]),
+            "password" => PasswordRules::rules([
+                "confirmed"
+            ]),
+            "name" => NameRules::rules()
+        ];
+    }
+}
