+++ app/Http/Requests/User/Rules/EmailRules.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Brewmap\Http\Requests\User\Rules;
+
+use Brewmap\Http\Requests\BaseRules;
+
+class EmailRules extends BaseRules 
+{
+    protected static array $rules = [
+        "required",
+        "email",
+    ];
+}
